removed synchronous code from event loop
