+++ config/controller_config.php
@@ -16,7 +16,7 @@ $controllers = [
     'print' =>        ZK\Controllers\PrintTags::class,
     'rss' =>          ZK\Controllers\RSS::class,
     'sso' =>          ZK\Controllers\SSOLogin::class,
-    'push' =>         ZK\Controllers\PushServer::class,
+    'push' =>         ZK\Service\PushServer::class,
     'validate' =>     ZK\Controllers\Validate::class,
     'challenge' =>    ZK\Controllers\Challenge::class,
     'turnstile' =>    ZK\Controllers\Turnstile::class,
